old/jobprocessortf/requirements.txt
fastapi==0.115.7
fastapi-cli==0.0.7
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
uvicorn==0.34.0
google-cloud-aiplatform==1.79.0